refactor(cli): move protocol command groups out of core/cli.py
Moves the `cativity`, `meshtastic`, `lora` and `vhci` Click groups into
modules/protocols/cli/, leaving core/cli.py to import and register them.
core/cli.py: 2065 -> 1478 lines. Command bodies moved verbatim; the only
edit is the depth of the relative imports.

The groups land in modules/protocols/cli/<x>.py rather than the planned
modules/protocols/<x>/cli.py because reaching the latter forces Python to
execute the protocol package's own __init__.py first, and those are eager:
sx1262 pulls matplotlib+numpy (+1.14s on every `catnip --help`), meshtastic
requires the `meshtastic` library (ImportError takes down the whole CLI when
it is absent), and vhci pulls `fcntl`, which does not exist on Windows.
modules/protocols/__init__.py is already empty and the new cli/__init__.py
holds only a docstring, so the static import chain stays free. The deferred
in-function imports of the protocol implementations are kept deferred.

Also:
- drop the now-unused `queue` import from core/cli.py
- widen the section 2.3 anti-cycle test, which globbed for files named
  cli.py and so would not have covered the new modules

Verified: no diff against tests/snapshots/cli_tree_linux.txt; no protocol
package in sys.modules after importing the CLI; pytest 325 passed / 6 failed,
identical to a worktree at HEAD (pre-existing bridge and missing-`usb`
failures).
