+++ catnip/modules/core/cli.py
@@ -10,7 +10,6 @@
 import logging
 import os
 import sys
-import queue
 
 # Internal
 from ..utils._version import __version__
@@ -28,6 +27,12 @@ from .catnip import (
     catnip_get_devices,
 )
 from .usb_connection import ShellConnection, CATSNIFFER_VID, CATSNIFFER_PID
+
+# Command groups assembled by build_cli()
+from ..protocols.cli.cativity import cativity as _cativity
+from ..protocols.cli.meshtastic import meshtastic as _meshtastic
+from ..protocols.cli.sx1262 import lora as _lora
+from ..protocols.cli.vhci import vhci as _vhci
 
 # External
 import click
@@ -1015,598 +1020,6 @@ def verify(test_all, device, quiet):
         sys.exit(1)
 
 
-@click.command()
-@click.option(
-    "--device",
-    "-d",
-    default=None,
-    type=int,
-    help="Device ID (for multiple CatSniffers)",
-)
-@click.option(
-    "--channel", "-c", type=click.IntRange(11, 26), help="Fixed channel (11-26)"
-)
-@click.option("--topology", "-t", is_flag=True, help="Show network topology")
-@click.option(
-    "--protocol",
-    "-p",
-    default="all",
-    type=click.Choice(["all", "zigbee", "thread"]),
-    help="Protocol filter",
-)
-def cativity(device, channel, topology, protocol):
-    """IQ Activity Monitor"""
-    from ..protocols.cativity.runner import CativityRunner
-
-    dev = get_device_or_exit(device)
-    cat = Catnip(dev.bridge_port)
-
-    # Verify firmware
-    print_info("Checking for Sniffer TI firmware...")
-    if cat.check_firmware_by_metadata("ti_sniffer", dev.shell_port):
-        print_success("Sniffer TI firmware found (via metadata)!")
-    elif cat.check_ti_firmware():
-        print_success("Sniffer TI firmware found (via direct communication)!")
-    else:
-        print_warning("Sniffer TI firmware not found! - Flashing Sniffer TI")
-        # Initialize Flasher for flashing
-        flasher_flash = Flasher()
-        if not flasher_flash.find_flash_firmware("ti_sniffer", dev):
-            print_error("Failed to flash Sniffer TI firmware")
-            return
-
-        print_info("Waiting for device to initialize...")
-        time.sleep(0.5)
-
-    # Send identification command to help identify which device was flashed
-    send_identify_command(dev)
-
-    print_info(f"[{dev}] Starting Cativity analysis...")
-    runner = CativityRunner(dev, console=console)
-    runner.run(channel=channel, topology=topology, protocol=protocol)
-
-
-# ===================== Meshtastic Commands =====================
-
-
-@click.group(context_settings={"help_option_names": ["-h", "--help"]})
-def meshtastic():
-    """Meshtastic protocol tools"""
-    pass
-
-
-@meshtastic.command("decode")
-@click.option(
-    "-i",
-    "--input",
-    required=True,
-    help="Hex-encoded payload (raw packet data starting with dest, sender, etc.)",
-)
-@click.option(
-    "-k",
-    "--key",
-    default="1PG7OiApB1nwvP+rz05pAQ==",
-    help="Base64-encoded AES key. Use 'ham' or 'nokey' for open channels",
-)
-def meshtastic_decode(input, key):
-    """Decrypt and decode a hex-encoded Meshtastic packet"""
-    try:
-        from ..protocols.meshtastic import MeshtasticDecoder
-    except ImportError as e:
-        print_error(
-            f"The 'meshtastic' library is required for this command. (Error: {e})"
-        )
-        print_empty_line()
-        print_warning("This library should be bundled with the package.")
-        print_info("If it's missing, you can install it manually:")
-        print_dim("pip install meshtastic protobuf pyyaml")
-        sys.exit(1)
-
-    try:
-        decoder = MeshtasticDecoder(key=key)
-        decrypted_hex, result = decoder.decode(input)
-        print(f"Decrypted raw (hex): {decrypted_hex}")
-        print(result)
-    except Exception as e:
-        print_error(f"Error: {e}")
-        sys.exit(1)
-
-
-@meshtastic.command("live")
-@click.option(
-    "-d",
-    "--device",
-    default=None,
-    type=int,
-    help="Device ID (for multiple CatSniffers)",
-)
-@click.option(
-    "-baud",
-    "--baudrate",
-    type=int,
-    default=115200,
-    help="Baudrate (default: 115200)",
-)
-@click.option(
-    "-f",
-    "--frequency",
-    type=float,
-    default=906.875,
-    help="Frequency in MHz (default: 906.875)",
-)
-@click.option(
-    "-ps",
-    "--preset",
-    type=click.Choice(
-        [
-            "defcon33",
-            "ShortTurbo",
-            "ShortSlow",
-            "ShortFast",
-            "MediumSlow",
-            "MediumFast",
-            "LongSlow",
-            "LongFast",
-            "LongMod",
-            "VLongSlow",
-        ]
-    ),
-    default="LongFast",
-    help="Channel preset (default: LongFast)",
-)
-def meshtastic_live(device, baudrate, frequency, preset):
-    """Live Meshtastic decoder - Capture and decode packets in real-time"""
-    try:
-        from ..protocols.meshtastic import MeshtasticLiveDecoder
-    except ImportError as e:
-        print_error(
-            f"The 'meshtastic' library is required for this command. (Error: {e})"
-        )
-        print_empty_line()
-        print_warning("This library should be bundled with the package.")
-        print_info("If it's missing, you can install it manually:")
-        print_dim("pip install meshtastic protobuf pyyaml")
-        sys.exit(1)
-
-    # Get device or exit with error
-    dev = get_device_or_exit(device)
-
-    # Use the LoRa port from the device
-    port = dev.lora_port
-    if not port:
-        print_error("LoRa port not found for device!")
-        return
-
-    # Use the Shell port for configuration
-    shell_port = dev.shell_port
-    if not shell_port:
-        print_error("Shell port not found for device! Required for configuration.")
-        return
-
-    decoder = MeshtasticLiveDecoder(port, baudrate)
-
-    freq_hz = int(frequency * 1_000_000)
-    print_info(f"Using device: {dev}")
-    print_info(f"Configuring radio: {frequency} MHz ({freq_hz} Hz), preset: {preset}")
-
-    # Configure radio using shell port with correct commands
-    if not decoder.configure_radio(freq_hz, preset, shell_port):
-        print_error("Failed to configure radio")
-        return
-
-    print_info("Starting capture... Press Ctrl+C to stop")
-    decoder.start()
-
-    try:
-        decoder.process_packets()
-    except KeyboardInterrupt:
-        print_info("Shutting down...")
-    finally:
-        decoder.stop()
-
-
-@meshtastic.command("dashboard")
-@click.option(
-    "-d",
-    "--device",
-    default=None,
-    type=int,
-    help="Device ID (for multiple CatSniffers)",
-)
-@click.option(
-    "-baud",
-    "--baudrate",
-    type=int,
-    default=115200,
-    help="Baudrate (default: 115200)",
-)
-@click.option(
-    "-f",
-    "--frequency",
-    type=float,
-    default=906.875,
-    help="Frequency in MHz (default: 906.875)",
-)
-@click.option(
-    "-ps",
-    "--preset",
-    type=click.Choice(
-        [
-            "defcon33",
-            "ShortTurbo",
-            "ShortSlow",
-            "ShortFast",
-            "MediumSlow",
-            "MediumFast",
-            "LongSlow",
-            "LongFast",
-            "LongMod",
-            "VLongSlow",
-        ]
-    ),
-    default="LongFast",
-    help="Channel preset (default: LongFast)",
-)
-def meshtastic_dashboard(device, baudrate, frequency, preset):
-    """Meshtastic Chat TUI - Beautiful terminal dashboard for Meshtastic"""
-    import asyncio
-
-    try:
-        from ..protocols.meshtastic.core import configure_meshtastic_radio
-        from ..protocols.meshtastic import MeshtasticChatApp, Monitor
-    except ImportError as e:
-        print_error(
-            f"The 'meshtastic' library is required for this command. (Error: {e})"
-        )
-        print_empty_line()
-        print_warning("This library should be bundled with the package.")
-        print_info("If it's missing, you can install it manually:")
-        print_dim("pip install meshtastic protobuf pyyaml")
-        sys.exit(1)
-
-    # Get device or exit with error
-    dev = get_device_or_exit(device)
-
-    # Use the LoRa port from the device
-    port = dev.lora_port
-    if not port:
-        print_error("LoRa port not found for device!")
-        return
-
-    # Use the Shell port for configuration
-    shell_port = dev.shell_port
-    if not shell_port:
-        print_error("Shell port not found for device! Required for configuration.")
-        return
-
-    print_info(f"Using device: {dev}")
-
-    # Create monitor
-    rx_queue = queue.Queue()
-    mon = Monitor(port, baudrate, rx_queue)
-    mon.start()
-
-    # Configure radio using shell port securely
-    print_info("Configuring radio...")
-    freq_hz = int(frequency * 1_000_000)
-
-    if not configure_meshtastic_radio(shell_port, freq_hz, preset):
-        print_error("Failed to configure radio")
-        mon.stop()
-        return
-
-    try:
-        app = MeshtasticChatApp(monitor=mon, preset=preset, freq=str(frequency))
-        asyncio.run(app.run_async())
-    finally:
-        mon.stop()
-
-
-@meshtastic.command("config")
-@click.argument("file")
-def meshtastic_config(file):
-    """Extract PSKs and config info from a Meshtastic JSONC config file"""
-    try:
-        from ..protocols.meshtastic import MeshtasticConfigExtractor
-    except ImportError as e:
-        print_error(
-            f"The 'meshtastic' library is required for this command. (Error: {e})"
-        )
-        print_empty_line()
-        print_warning("This library should be bundled with the package.")
-        print_info("If it's missing, you can install it manually:")
-        print_dim("pip install meshtastic protobuf pyyaml")
-        sys.exit(1)
-
-    extractor = MeshtasticConfigExtractor(file)
-    if extractor.load():
-        extractor.print_all()
-    else:
-        sys.exit(1)
-
-
-# ===================== LoRa Commands =====================
-
-
-@click.group(context_settings={"help_option_names": ["-h", "--help"]})
-def lora():
-    """LoRa SX1262 tools"""
-    pass
-
-
-@lora.command("spectrum")
-@click.option(
-    "-d",
-    "--device",
-    default=None,
-    type=int,
-    help="Device ID (for multiple CatSniffers)",
-)
-@click.option(
-    "-b",
-    "--baudrate",
-    type=int,
-    default=115200,
-    help="Baudrate (default: 115200)",
-)
-@click.option(
-    "--start-freq",
-    type=float,
-    default=150,
-    help="Starting frequency in MHz (default: 150)",
-)
-@click.option(
-    "--end-freq",
-    type=float,
-    default=960,
-    help="End frequency in MHz (default: 960)",
-)
-@click.option(
-    "--offset",
-    type=int,
-    default=-15,
-    help="RSSI offset in dBm (default: -15)",
-)
-def lora_spectrum(device, baudrate, start_freq, end_freq, offset):
-    """Live Spectrum Scanner for SX1262 - Real-time frequency spectrum analyzer"""
-    from ..protocols.sx1262.spectrum import SpectrumScan
-
-    # Get device or exit with error
-    dev = get_device_or_exit(device)
-
-    # The spectral scan is driven by the RP2040 text shell (CDC2), not the
-    # LoRa data stream: it sends set_start_freq/set_end_freq/start and reads
-    # back the FREQ/SCAN frames on that same port.
-    port = dev.shell_port
-    if not port:
-        print_error("Shell port not found for device! Required for spectrum scan.")
-        return
-
-    print_info(f"Using device: {dev}")
-    print_info(f"Starting spectrum scan: {start_freq}-{end_freq} MHz")
-
-    scanner = SpectrumScan(port=port, baudrate=baudrate)
-
-    try:
-        scanner.run(start_freq=start_freq, end_freq=end_freq, rssi_offset=offset)
-    except KeyboardInterrupt:
-        scanner.stop_task()
-
-
-# ===================== VHCI Bridge Commands =====================
-
-
-@click.group(context_settings={"help_option_names": ["-h", "--help"]})
-def vhci():
-    """VHCI Bridge - Expose CatSniffer as hciX.
-
-    Requires sudo and the hci_vhci kernel module.
-
-    \b
-        sudo modprobe hci_vhci
-        sudo python3 catnip.py vhci start
-    """
-    pass
-
-
-@vhci.command("start")
-@click.option(
-    "--device",
-    "-d",
-    default=None,
-    type=int,
-    help="Device ID (for multiple CatSniffers)",
-)
-@click.option(
-    "--baud",
-    default=2000000,
-    type=int,
-    show_default=True,
-    help="Baud rate for serial port",
-)
-@click.option("--verbose", "-v", is_flag=True, help="Enable verbose (DEBUG) logging")
-def vhci_start(device, baud, verbose):
-    """Start the VHCI bridge — CatSniffer appears as hciX.
-
-    Requires root privileges and the hci_vhci kernel module:
-
-    \b
-        sudo modprobe hci_vhci
-        sudo catnip vhci start
-        hciconfig -a
-
-    Compatible tools: bluetoothctl, btmgmt, btmon, bleak, bettercap.
-    """
-    import signal
-    from ..protocols.vhci import VHCIBridge
-
-    if os.geteuid() != 0 and not os.access("/dev/vhci", os.R_OK | os.W_OK):
-        print_warning(
-            "Insufficient permissions for /dev/vhci access. Try running with sudo or check group membership."
-        )
-
-    if not os.path.exists("/dev/vhci"):
-        print_error("/dev/vhci not found. Load the kernel module first:")
-        print_dim("sudo modprobe hci_vhci")
-        sys.exit(1)
-
-    # Resolve device
-    if device is not None:
-        dev = catnip_get_device(device)
-        if dev is None:
-            print_error(f"CatSniffer device #{device} not found.")
-            sys.exit(1)
-        if not dev.bridge_port:
-            print_error(f"Device #{device} has no Cat-Bridge port detected.")
-            sys.exit(1)
-        print_info(f"Using device {dev}, port: {dev.bridge_port}")
-    else:
-        devs = catnip_get_devices()
-        if devs and devs[0].bridge_port:
-            dev = devs[0]
-            print_info(f"Auto-detected CatSniffer: {dev}, port: {dev.bridge_port}")
-        else:
-            print_error("CatSniffer not found. Connect a device or specify -d.")
-            sys.exit(1)
-
-    # Firmware check
-    cat = Catnip(dev.bridge_port)
-    print_info("Checking for Sniffle firmware...")
-    if cat.check_firmware_by_metadata("sniffle", dev.shell_port):
-        print_success("Sniffle firmware found!")
-    else:
-        print_warning("Sniffle firmware not found — flashing now...")
-        flasher = Flasher()
-        if not flasher.find_flash_firmware("sniffle", dev):
-            print_error("Failed to flash Sniffle firmware. Aborting.")
-            sys.exit(1)
-        print_info("Waiting for device to initialize...")
-        time.sleep(1)
-        if cat.check_firmware_by_metadata("sniffle", dev.shell_port):
-            print_success("Sniffle firmware verified!")
-        else:
-            print_warning("Firmware verification failed, continuing anyway...")
-
-    # Logging
-    level = logging.DEBUG if verbose else logging.INFO
-    logging.basicConfig(
-        level=level,
-        format="%(message)s",
-        handlers=[RichHandler(console=console, show_time=False)],
-    )
-    log = logging.getLogger("vhci")
-
-    bridge = VHCIBridge(dev.bridge_port, log)
-
-    def _shutdown(sig, frame):
-        print_warning("Shutting down VHCI bridge...")
-        bridge.stop()
-        sys.exit(0)
-
-    signal.signal(signal.SIGINT, _shutdown)
-    signal.signal(signal.SIGTERM, _shutdown)
-
-    try:
-        bridge.start()
-    except Exception as e:
-        print_error(f"Failed to start bridge: {e}")
-        sys.exit(1)
-
-    print_success("Bridge running. Device should appear as hciX.")
-    print_dim("Check with: hciconfig -a   |   Press Ctrl+C to stop")
-
-    try:
-        bridge.run()
-    except KeyboardInterrupt:
-        pass
-    finally:
-        bridge.stop()
-
-
-@vhci.command("check")
-def vhci_check():
-    """Check VHCI bridge prerequisites (kernel module, /dev/vhci, root, packages)."""
-    import subprocess as _sp
-
-    all_ok = True
-
-    # Permissions check
-    if os.access("/dev/vhci", os.R_OK | os.W_OK):
-        print_success("  permissions  : OK (access to /dev/vhci)")
-    elif os.geteuid() == 0:
-        print_success("  root         : OK")
-    else:
-        print_warning(
-            "  permissions  : Insufficient — bridge may fail to open /dev/vhci"
-        )
-        all_ok = False
-
-    # Kernel module
-    try:
-        result = _sp.run(["lsmod"], capture_output=True, text=True, timeout=5)
-        if "hci_vhci" in result.stdout:
-            print_success("  hci_vhci     : loaded")
-        else:
-            print_warning("  hci_vhci     : NOT loaded — run: sudo modprobe hci_vhci")
-            all_ok = False
-    except Exception:
-        print_error("  hci_vhci     : could not run lsmod")
-        all_ok = False
-
-    # /dev/vhci
-    if os.path.exists("/dev/vhci"):
-        print_success("  /dev/vhci    : exists")
-    else:
-        print_warning("  /dev/vhci    : missing — run: sudo modprobe hci_vhci")
-        all_ok = False
-
-    # BlueZ (bluetoothctl)
-    try:
-        _sp.run(["bluetoothctl", "--version"], capture_output=True, timeout=3)
-        print_success("  bluetoothctl : found")
-    except FileNotFoundError:
-        print_warning("  bluetoothctl : not found — install bluez")
-        all_ok = False
-    except Exception:
-        print_warning("  bluetoothctl : check failed")
-
-    # btmon
-    try:
-        _sp.run(["btmon", "--version"], capture_output=True, timeout=3)
-        print_success("  btmon        : found")
-    except FileNotFoundError:
-        print_dim("  btmon        : not found (optional — install bluez-utils)")
-    except Exception:
-        pass
-
-    # bleak (Python)
-    try:
-        import bleak  # noqa: F401
-
-        print_success("  bleak        : installed")
-    except ImportError:
-        print_dim("  bleak        : not installed (optional — pip install bleak)")
-
-    # CatSniffer device
-    devs = catnip_get_devices()
-    if devs:
-        for dev in devs:
-            port = dev.bridge_port or "?"
-            print_success(f"  CatSniffer   : {dev}  bridge={port}")
-    else:
-        print_warning("  CatSniffer   : no device detected")
-        all_ok = False
-
-    print_empty_line()
-    if all_ok:
-        if os.access("/dev/vhci", os.R_OK | os.W_OK):
-            print_success("All prerequisites met. Run: catnip vhci start")
-        else:
-            print_success("All prerequisites met. Run: sudo catnip vhci start")
-    else:
-        print_warning("Some prerequisites are missing. See above.")
-
-
 # ===================== Firmware Update Commands =====================
 
 
@@ -2045,12 +1458,12 @@ def build_cli() -> click.Group:
     without running it (tests, snapshot dumps).
     """
     cli.add_command(sniff)
-    cli.add_command(cativity)
-    cli.add_command(meshtastic)
+    cli.add_command(_cativity)
+    cli.add_command(_meshtastic)
     cli.add_command(restore)
-    cli.add_command(lora)
+    cli.add_command(_lora)
     if platform.system() == "Linux":
-        cli.add_command(vhci)
+        cli.add_command(_vhci)
         cli.add_command(setup_env)
     cli.add_command(verify)
     if platform.system() in ["Linux", "Darwin"]:
